int -> size_t v2

+++ modules/task_1/kamskov_e_shell_batcher_sort/shell_batcher.cpp
@@ -43,7 +43,7 @@ std::vector<int> batchShellSort(const std::vector<int>& v, int size) {
 std::vector<int> evenBatch(const std::vector<int>& A, const std::vector<int>& B) {
     int size_res = A.size() / 2 + B.size() / 2 + A.size() % 2 + B.size() % 2;
     std::vector <int> res(size_res);
-    int iA = 0, iB = 0, i = 0;
+    long unsigned int iA = 0, iB = 0, i = 0;
     while ((iA < A.size()) && (iB < B.size())) {
         if (A[iA] <= B[iB]) {
             res[i] = A[iA];
@@ -55,12 +55,12 @@ std::vector<int> evenBatch(const std::vector<int>& A, const std::vector<int>& B)
         ++i;
     }
     if (iA >= A.size()) {
-        for (int j = iB; j < B.size(); j = j + 2) {
+        for (long unsigned int j = iB; j < B.size(); j = j + 2) {
             res[i] = B[j];
             ++i;
         }
     } else {
-        for (int j = iA; j < A.size(); j = j + 2) {
+        for (long unsigned int j = iA; j < A.size(); j = j + 2) {
             res[i] = A[j];
             ++i;
         }
@@ -73,7 +73,7 @@ std::vector<int> oddBatch(const std::vector<int>& A, const std::vector<int>& B) 
     int sizeB = B.size();
     int size_res = sizeA / 2 + sizeB / 2;
     std::vector <int> res(size_res);
-    int iA = 1, iB = 1, i = 0;
+    size_t iA = 1, iB = 1, i = 0;
     while ((iA < sizeA) && (iB < sizeB)) {
         if (A[iA] <= B[iB]) {
             res[i] = A[iA];
@@ -85,12 +85,12 @@ std::vector<int> oddBatch(const std::vector<int>& A, const std::vector<int>& B) 
         ++i;
     }
     if (iA >= sizeA) {
-        for (int j = iB; j < sizeB; j = j + 2) {
+        for (size_t j = iB; j < sizeB; j = j + 2) {
             res[i] = B[j];
             ++i;
         }
     } else {
-        for (int j = iA; j < sizeA; j = j + 2) {
+        for (size_t j = iA; j < sizeA; j = j + 2) {
             res[i] = A[j];
             ++i;
         }
@@ -126,6 +126,6 @@ std::vector<int> mergeBatch(const std::vector<int>& even, const std::vector<int>
 }
 
 void print(std::vector<int> const& a) {
-    for (int i = 0; i < a.size(); ++i) std::cout << a[i] << ' ';
+    for (size_t i = 0; i < a.size(); ++i) std::cout << a[i] << ' ';
     std::cout << std::endl;
 }
